refactor: move todo gateway impls to its own package
