+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/config/TodoGatewayConfig.java
@@ -1,7 +1,12 @@
 package br.com.jpcchaves.infrastructure.config;
 
 import br.com.jpcchaves.application.usecaseimpl.*;
-import br.com.jpcchaves.infrastructure.gatewayimpl.*;
+import br.com.jpcchaves.infrastructure.gatewayimpl.todo.CreateTodoGatewayImpl;
+import br.com.jpcchaves.infrastructure.gatewayimpl.todo.DeleteTodoGatewayImpl;
+import br.com.jpcchaves.infrastructure.gatewayimpl.todo.GetTodoByIdGatewayImpl;
+import br.com.jpcchaves.infrastructure.gatewayimpl.todo.ListTodoGatewayImpl;
+import br.com.jpcchaves.infrastructure.gatewayimpl.todo.UpdateTodoGatewayImpl;
+import br.com.jpcchaves.infrastructure.gatewayimpl.todo.UpdateTodoStatusGatewayImpl;
 import br.com.jpcchaves.usecase.todo.CreateTodoUseCase;
 import br.com.jpcchaves.usecase.todo.DeleteTodoUseCase;
 import br.com.jpcchaves.usecase.todo.GetTodoByIdUseCase;
