+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/gatewayimpl/todo/GetTodoByIdGatewayImpl.java
@@ -1,32 +1,37 @@
-package br.com.jpcchaves.infrastructure.gatewayimpl;
+package br.com.jpcchaves.infrastructure.gatewayimpl.todo;
 
-import br.com.jpcchaves.application.gateway.todo.UpdateTodoStatusGateway;
-import br.com.jpcchaves.core.domain.enums.TodoStatus;
+import br.com.jpcchaves.application.gateway.todo.GetTodoGateway;
+import br.com.jpcchaves.core.domain.Todo;
 import br.com.jpcchaves.core.exception.TodoException;
 import br.com.jpcchaves.core.exception.enums.ExceptionDefinition;
+import br.com.jpcchaves.infrastructure.mapper.TodoMapper;
 import br.com.jpcchaves.infrastructure.persistence.entity.TodoEntity;
 import br.com.jpcchaves.infrastructure.persistence.repository.IRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 @Component
-public class UpdateTodoStatusGatewayImpl implements UpdateTodoStatusGateway {
+public class GetTodoByIdGatewayImpl implements GetTodoGateway {
   @Qualifier("${deploy.repo}")
   private final IRepository<TodoEntity, Long> todoRepository;
 
-  public UpdateTodoStatusGatewayImpl(IRepository<TodoEntity, Long> todoRepository) {
+  private final TodoMapper todoMapper;
+
+  public GetTodoByIdGatewayImpl(
+      IRepository<TodoEntity, Long> todoRepository, TodoMapper todoMapper) {
     this.todoRepository = todoRepository;
+    this.todoMapper = todoMapper;
   }
 
   @Override
-  public void updateStatus(Long id, TodoStatus status) {
+  @Transactional(readOnly = true)
+  public Todo getTodo(Long id) {
     TodoEntity todoEntity =
         todoRepository
             .findById(id)
             .orElseThrow(() -> new TodoException(ExceptionDefinition.TD0001));
 
-    todoEntity.setStatus(status);
-
-    todoRepository.save(todoEntity);
+    return todoMapper.toTodo(todoEntity);
   }
 }
