+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/gatewayimpl/todo/ListTodoGatewayImpl.java
@@ -1,37 +1,30 @@
-package br.com.jpcchaves.infrastructure.gatewayimpl;
+package br.com.jpcchaves.infrastructure.gatewayimpl.todo;
 
-import br.com.jpcchaves.application.gateway.todo.GetTodoGateway;
+import br.com.jpcchaves.application.gateway.todo.ListTodoGateway;
 import br.com.jpcchaves.core.domain.Todo;
-import br.com.jpcchaves.core.exception.TodoException;
-import br.com.jpcchaves.core.exception.enums.ExceptionDefinition;
 import br.com.jpcchaves.infrastructure.mapper.TodoMapper;
 import br.com.jpcchaves.infrastructure.persistence.entity.TodoEntity;
 import br.com.jpcchaves.infrastructure.persistence.repository.IRepository;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 @Component
-public class GetTodoByIdGatewayImpl implements GetTodoGateway {
+public class ListTodoGatewayImpl implements ListTodoGateway {
   @Qualifier("${deploy.repo}")
   private final IRepository<TodoEntity, Long> todoRepository;
 
   private final TodoMapper todoMapper;
 
-  public GetTodoByIdGatewayImpl(
-      IRepository<TodoEntity, Long> todoRepository, TodoMapper todoMapper) {
+  public ListTodoGatewayImpl(IRepository<TodoEntity, Long> todoRepository, TodoMapper todoMapper) {
     this.todoRepository = todoRepository;
     this.todoMapper = todoMapper;
   }
 
   @Override
   @Transactional(readOnly = true)
-  public Todo getTodo(Long id) {
-    TodoEntity todoEntity =
-        todoRepository
-            .findById(id)
-            .orElseThrow(() -> new TodoException(ExceptionDefinition.TD0001));
-
-    return todoMapper.toTodo(todoEntity);
+  public List<Todo> list() {
+    return todoMapper.toTodoList(todoRepository.findAll());
   }
 }
